+++ src/components/EventFrom.js
@@ -44,7 +44,7 @@ const EventFrom = () => {
           </div>
 
           <button className="btn btn-primary" onClick={addEvent} disabled={unCreatable}>イベントを作成する</button>
-          <button className="btn btn-danger" onClick={deleteAllEvents} disabled={state.length === 0}>全てのイベントを削除する</button>
+          <button className="btn btn-danger" onClick={deleteAllEvents} disabled={state.events.length === 0}>全てのイベントを削除する</button>
         </form>
     </>
   )
