+++ src/components/Events.js
@@ -17,7 +17,7 @@ const Events = () => {
             </tr>
           </thead>
           <tbody>
-            {state.map((event, index) => (<Event key={index} event={event}/>))}
+            {state.events.map((event, index) => (<Event key={index} event={event}/>))}
           </tbody>
         </table>
     </>
